span:edition, span:version, span:note: https://github.com/metanorma/mn-samples-bsi/issues/10
